refactor: move spectrum_utils.py to core/data/spectrum_processing.py
