+++ simba/core/data/loaders.py
@@ -10,9 +10,9 @@ from simba.core.chemistry import chem_utils
 from simba.core.data.nist_loader import NistLoader
 from simba.core.data.preprocessing import PreprocessingUtils
 from simba.core.data.spectrum import SpectrumExt
+from simba.core.data.spectrum_processing import spectrum_hash
 from simba.logger_setup import logger
 from simba.murcko_scaffold import MurckoScaffold
-from simba.spectrum_utils import spectrum_hash
 
 
 class LoadData:
